requirements.txt
openai
poe-api
GoogleBard
EdgeGPT